fix: protect feature from user

+++ controllers/bookController.js
@@ -2,29 +2,52 @@ const response = require("../helper/response");
 const db = require("../models");
 
 const Book = db.book;
+const User = db.user;
 const addNewBook = async (req, res) => {
   try {
-    let info = {
-      title: req.body.title,
-      author: req.body.author,
-    };
+    let user = await User.findOne({
+      where: { code_refferal: req.body.code_refferal },
+    });
+    if (user.code_refferal === "NAMTHIP") {
+      let info = {
+        title: req.body.title,
+        author: req.body.author,
+        quantity: req.body.quantity,
+      };
 
-    const product = await Book.create(info);
-    res.send(response(200, product, "Success created new book"));
-    console.log(product);
+      const product = await Book.create(info);
+
+      return res.send(response(200, product, "Success created new book"));
+    } else {
+      return res.send(
+        response(400, null, "Cannot access because youre not admin")
+      );
+    }
   } catch (error) {
     throw error;
   }
 };
 
 const getOneBook = async (req, res) => {
-  let id = req.params.id;
-  let product = await Book.findOne({ where: { id: id } });
-  res.send(response(200, product, "Success get book"));
+  try {
+    let id = req.params.id;
+    let product = await Book.findOne({ where: { id: id } });
+    res.send(response(200, product, "Success get book"));
+  } catch (error) {
+    console.log(error);
+  }
 };
 
 const updateDataBook = async (req, res) => {
   try {
+    let user = await User.findOne({
+      where: { code_refferal: req.body.code_refferal },
+    });
+    if (user.code_refferal !== "NAMTHIP") {
+      return res.send(
+        response(400, null, "Cannot access because youre not admin")
+      );
+    }
     let id = req.params.id;
 
     await Book.update(req.body, { where: { id: id } });
@@ -37,6 +60,14 @@ const updateDataBook = async (req, res) => {
 
 const deleteDataBook = async (req, res) => {
   try {
+    let user = await User.findOne({
+      where: { code_refferal: req.body.code_refferal },
+    });
+    if (user.code_refferal !== "NAMTHIP") {
+      return res.send(
+        response(400, null, "Cannot access because youre not admin")
+      );
+    }
     let id = req.params.id;
     let products = await Book.findAll({});
     await Book.destroy({ where: { id: id } });
